Encapsulate sprint creation methods in helper file
