+++ packages/frontend/src/pages/projectPages/ProjectSprints.tsx
@@ -2,14 +2,12 @@ import React, { useMemo, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { List, message, Typography } from 'antd';
 import dayjs from 'dayjs';
-import type { Dayjs } from 'dayjs';
 import { DragDropContext, DropResult } from 'react-beautiful-dnd';
 import { useGetSprintsByProjectIdQuery } from '../../api/sprint';
 import { useUpdateBacklogMutation } from '../../api/socket/backlogHooks';
 import type { Sprint } from '../../api/sprint';
 import SprintCreationModal from '../../components/modals/SprintCreationModal';
 import SprintListingCard from '../../components/cards/SprintListingCard';
-import type { AutoSprintTypes } from '../../helpers/SprintModal.types';
 import BacklogList from '../../components/lists/BacklogList';
 import BacklogCreationModal from '../../components/modals/BacklogCreationModal';
 import StrictModeDroppable from '../../components/dnd/StrictModeDroppable';
@@ -19,50 +17,7 @@ import EpicCreationModal from '../../components/modals/EpicCreationModal';
 import { Epic } from '../../api/types';
 import EpicListingCard from '../../components/cards/EpicListingCard';
 import { useGetEpicsByProjectIdQuery } from '../../api/backlog';
-
-export const GENERIC_NEW_SPRINT: AutoSprintTypes = {
-  name: 'Sprint 1',
-  dates: [dayjs(), dayjs().add(7, 'day')],
-  duration: 1,
-};
-
-const incrementSprintName = (latestSprintName: string): string => {
-  // Get last digit of the name
-  const nameTokens: string[] = latestSprintName.split(' ');
-  let sprintCount: number = +nameTokens[nameTokens.length - 1];
-  if (isNaN(sprintCount)) {
-    sprintCount = 2;
-  } else {
-    // Remove current count from the name if it exists
-    nameTokens.pop();
-  }
-  // Concat latest count to the name
-  return `${nameTokens.join(' ')} ${sprintCount + 1}`;
-};
-
-const incrementSprintDate = (duration: number, dates: string[]): Dayjs[] | undefined => {
-  if (!dates[0] || !dates[1]) {
-    return undefined;
-  }
-  if (duration === 0) {
-    const startDate = dayjs(dates[0]);
-    const endDate = dayjs(dates[1]);
-    const currentRange = endDate.diff(startDate, 'day');
-    return [endDate, endDate.add(currentRange, 'day')];
-  }
-  return [dayjs(dates[1]), dayjs(dates[1]).add(duration * 7, 'day')];
-};
-
-export const autoSuggestNewSprint = (latestSprint: Sprint): AutoSprintTypes => {
-  // Only works if sprint follows a naming convention with a digit at the back (e.g Sprint 1 or Week 1)
-  const newSprintName = incrementSprintName(latestSprint.name);
-  const newDates = incrementSprintDate(latestSprint.duration, [latestSprint.start_date, latestSprint.end_date]);
-  return {
-    name: newSprintName,
-    dates: newDates,
-    duration: latestSprint.duration,
-  };
-};
+import { GENERIC_NEW_SPRINT, autoSuggestNewSprint } from '../../helpers/sprintCreationHelper';
 
 function ProjectSprints(): JSX.Element {
   const params = useParams();
